+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## Master
-* Improve stream format error [#950](https://github.com/membraneframework/membrane_core/pull/950)
+ * Improve stream format error. [#950](https://github.com/membraneframework/membrane_core/pull/950)
+ * Minor fixes in `Membrane.Connector`. [#952](https://github.com/membraneframework/membrane_core/pull/952)
 
 ## 1.2.0
  * Add `:max_instances` option for dynamic pads. [#876](https://github.com/membraneframework/membrane_core/pull/876)